m_batt_meas: correct interpretation of XC6804 CSO signal
The XC6804 charging status output pin is on during trickle and main
charging, and off when charging is complete.  The sense of the signal
was inverted in the code, resulting in a CHARGING_FINISHED event when
the USB is connected and the battery is partially discharged.  Swap the
interpretation so a CHARGING event is generated instead.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>